cortex: persist typelib info into the cortex
this is a one-way backing only.
types added to the typelib are added as 'syn:type' tufos, which can be inspected via the cortex.
however, changes made via the cortex/swarm *do not* affect the installed types.
likewise, upon reload of the cortex, the types still must be re-installed.
the cortex persistence is just metadata.