datamodel: disallow mixed case props
the cortex authors *suggest* all prop names should be lowercase; however, this policy should be enforced, and data models should be rejected that contain mixed case prop names.

the current behavior is to accept mixed cased prop names, but implicitly lowercase them in the data model tufos. personally, i find this quite unexpected. it would be better to raise an exception and highlight the issue as early as possible.