Dendrite: File Parsing Sub-system
This is a first pass at building a file parsing subsystem in synapse. Dendrite
consists of 3 distributed components: Coordinator, Jobs, and Parser. It is
intended that specialized parsers inherit from Parser and provide MIME specific
parsing. Each Parser registers with the Coordinator, effectively directing MIME
specific jobs to be placed on certain named queues within the Jobs component. As
jobs are added to the queue, the Parsers are notified via the Service Bus and
then consume jobs off the queue and invoke the overridden process method within
its specialized instance.