The boss _onJobFini handler is race-prone - this addresses an issue with a ondone() function being executed AFTER a job waitevt is set().
This can be problematic if thread A calls a job and then waits for its completion, expecting work to be done in a ondone function as well.
The worker thread will set() that wait event, allowing a context switch back to thread A, which can be problematic if threadA was expecting
the ondone() function to be executed as part of the job completion.