Change Link/Chan from EventBus to s_coro.Fini.
Move closing the daemon item generator back into async loop because
(we now understand that) the close actually inline handles the generator
cleanup, and that's where the actual transaction commit/abort occurs.

Remove unnecessary next() after close for the item generator.

Added an assert into layer to make sure that lmdb write transaction is
committed/aborted on same thread as it was created on.