Add a unit test for lift issue, solved it.
The issue was that in a situation where the lower layer matched a
filter, but a higher layer changed a prop so that the filter didn't
match anymore would still return the node.

Solved by keeping track of the property name and the "originating"
layer, i.e. the layer that the query was currently being done to.