fix multi-layer lift results (#809)
The issue was that in a situation where the lower layer matched a
filter, but a higher layer changed a prop so that the filter didn't
match anymore would still return the node.

Solved by keeping track of the property name and the "originating"
layer, i.e. the layer that the query was currently being done to. This also 
fixes a related issue in which the same node could be returned twice
in a query if it was in more than 1 layer.

* Removed * import, unused imports and variables.

* Add an optional parameter to getTestCore to allow configuration.

* Fix formatting, unused variables.

* Add unit tests for lift issue, duplicate issue

* Moved stor* methods from layer to xact.

* Xact.buidcache stores dicts instead of list.

* Mov prop consolidation from Node._loadNodeData to snap._getNodeByBuid.

* Move layer-level logic to snap-getRowNodes

* Resolve issue where tag value wasn't being considered at all in getNodesBy.

* Made rawprop required for snap.getLiftNodes, et. al.

* Make sure getRowNodes is coherent with the snap buidcache.