Fix remote layer bug, reorganize cortex tests
Divide CortexTest into CortexTest and CortexBasicTest, so that
CortexTest can be inherited by tests with different layers and not have
superfluous tests.

Have RemoteLayerTest inherit from CortexTest again for better coverage
of cortexes with a two-layer setup with a remote layer.

Add a simple test to snap to directly test the bug found:  lifting by
primary property from a lower layer when an upper layer didn't have the
node improperly skipped it.