Implement merge method on view (#1505)
A new view that was forked from another view can now be merged back into the view it was forked on.

All changes from the child view are incorporated in its parent view and it is destroyed. As triggers were fired when they were applied to the child view, triggers do not fire when merged into the parent.

Limitation: a view can't merge into its parent if it has any existing forked children.