Cell.Ahainfo updates (SYN-4490, SYN-5194) (#3008)
Remove the Cell.ahaclient Telepath client from being cached in the synapse.lib.aha_clients dictionary. This allows the Cell Aha client to operate independently of a (potentially) globally cached client. This simplifies unit testing with multiple cells communicating with a single Aha service, as the client is now properly fini'd upon teardown of the cell. Since the AhaApi class maintains state for the service registration, this also means a user interacting with the global object directly cannot not affect the state of the service registration for a Cell.

Change the cell.ahainfo from being a static constant to being generated on demand by Cell.getAhaInfo(). This allows the cell to report the correct ready status to the Aha service, instead of a potentailly invalid static value. The previous behavior presented as having a incorrect ready status upon the Aha service restarting & service mirrors reporting the incorrect status.

When the cell mirror loop exits, it now reports its ready status to the Aha service. This allows a service to mark itself as "not ready" since exiting that loop may mean it is no longer in the realtime change window.

Fix a bug in nexus._tellAhaReady(self, status) which passed in a static value ( True ) to the Aha service instead of the the status value being passed in.