Convert one of the assertiosn in test_cell_minspace to use an asyncLoggerStream
This is a backport of some changes from c4c7a66a30c1b0b74a57c798f5d9ff3c9b83d703.
When merging this into synapse-3xx we will need to prefer this implemention since
it has a clearly await timeout + assertion that the message was seen in the log stream.