Convert deprecations from using Python warnings to log messages (SYN-9889) (#4582)
- [x] Converted deprecation functions ( s_common.deprecated /
s_common.deprdate ) to use the Python logging module instead of the
warnings module.
- [x] Updated devops documentation to reflect that the logging module
will no longer be used.
- [x] Updated several deprecation messages to make them more clear OR to
reduce them from being emitted during normal operation of the system.
The second case was because of several changes to the AHA APIs will be
transparent to users, and the deprecated functionality is really only
used in the multi-network case.
- [x] Removed the unused bruteVersionStr function from the inet
coremodule. The onset prop which used that in the past has also had its
logging updated to surpress BadTypeValu exceptions when attempting to
parse version values as semvers, since that message was misleading with
the changes made in 338679c5e9652bf61669f694f71d2dc5932deb90

---------

Co-authored-by: Cisphyx <cisphyx@vertex.link>
Co-authored-by: invisig0th <visi@vertex.link>